+++ src/i_input.h
@@ -25,7 +25,7 @@
 void I_OpenController(int which);
 void I_CloseController(int which);
 
-float I_AccelerateMouse(int val);
+double I_AccelerateMouse(int val);
 void I_ReadMouse(void);
 void I_UpdateJoystick(void);
